app/requirements/requirements.txt
# base modules
Django==1.9.6
djangorestframework==3.3.3
django_extensions==1.6.7
django_compressor==2.0
django-authtools==1.5.0
django-split-settings==0.1.3
django_rq==0.9.0

# pipeline
python-magic==0.4.11

# essentia
numpy==1.11.0

# grapher /
Pillow==3.2.0
scikits.audiolab==0.11.0
fmoo-audiotools==3.1b1